requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
python-dotenv
plotly
bcrypt
streamlit-javascript
PyJWT
